backend/requirements.txt
flask
requests
dnspython
python-dotenv
